bridge_cffi: route all results through BamlOutboundResult; module reorg (33b)
