+++ baml_language/crates/bridge_cffi/src/ffi/mod.rs
@@ -1,7 +1,6 @@
 //! FFI module - C-compatible entry points.
 
 pub mod callbacks;
-pub mod functions;
 pub mod handle;
 pub mod host_value;
 pub mod objects;
