+++ baml_language/sdks/go/bridge_go/pkg/errors.go
@@ -1,33 +1,18 @@
 package pkg
 
-import "strings"
-
-// BamlError is the base error type for all BAML errors.
+// BamlError is the single error type for all BAML errors surfaced to Go callers.
+//
+// The engine reports thrown values through a structured BamlOutboundResult
+// envelope (see decodeResult/thrownError), so the fully-qualified class name is
+// available directly on ClassName (e.g. "baml.errors.InvalidArgument",
+// "baml.panics.Cancelled", or a user error class). Callers that need to
+// discriminate match on ClassName rather than on a typed Go subtype — see 33b
+// for why the typed subtypes were removed.
 type BamlError struct {
+	// ClassName is the FQN of the thrown value's class, when available.
+	ClassName string
+	// Message is the rendered error string (class name, message, and traceback).
 	Message string
 }
 
 func (e *BamlError) Error() string { return e.Message }
-
-// BamlInvalidArgumentError indicates invalid input.
-type BamlInvalidArgumentError struct{ BamlError }
-
-// BamlClientError indicates an engine/runtime error.
-type BamlClientError struct{ BamlError }
-
-// BamlCancelledError indicates the operation was cancelled.
-type BamlCancelledError struct{ BamlError }
-
-// ParseBamlError parses a prefixed error string from Rust into a typed Go error.
-func ParseBamlError(msg string) error {
-	if strings.HasPrefix(msg, "BamlError: BamlInvalidArgumentError: ") {
-		return &BamlInvalidArgumentError{BamlError{Message: msg}}
-	}
-	if strings.HasPrefix(msg, "BamlError: BamlCancelledError: ") {
-		return &BamlCancelledError{BamlError{Message: msg}}
-	}
-	if strings.HasPrefix(msg, "BamlError: BamlClientError: ") {
-		return &BamlClientError{BamlError{Message: msg}}
-	}
-	return &BamlError{Message: msg}
-}
