+++ baml_language/sdks/nodejs/bridge_nodejs/src/runtime.rs
@@ -1,7 +1,7 @@
 //! BamlRuntime napi class.
 //!
 //! A zero-sized handle: the single source of truth for the `Arc<dyn Bex>`
-//! singleton is `bridge_cffi`, fetched via `bridge_cffi::engine::get_runtime()`
+//! singleton is `bridge_cffi`, fetched via `bridge_cffi::get_runtime()`
 //! at each call site (mirrors `bridge_python` after 31e-phase4), so this no
 //! longer caches its own clone.
 
@@ -32,7 +32,7 @@ impl BamlRuntime {
     ) -> napi::Result<Self> {
         // `initialize_runtime` stores the `Arc<dyn Bex>` in bridge_cffi's
         // singleton; we don't keep our own copy.
-        match bridge_cffi::engine::initialize_runtime(&root_path, files) {
+        match bridge_cffi::initialize_runtime(&root_path, files) {
             Ok(_bex) => Ok(BamlRuntime {}),
             Err(e) => Err(bridge_error_to_napi(e)),
         }
@@ -48,7 +48,7 @@ impl BamlRuntime {
         collectors: Option<Vec<&Collector>>,
         abort_controller: Option<&AbortController>,
     ) -> napi::Result<Buffer> {
-        let runtime = bridge_cffi::engine::get_runtime().map_err(bridge_error_to_napi)?;
+        let runtime = bridge_cffi::get_runtime().map_err(bridge_error_to_napi)?;
         let kwargs = decode_args(args_proto.as_ref(), &function_name)?;
         let host_ctx = ctx.and_then(|c| c.host_span_context());
         let cancel = abort_controller
@@ -70,7 +70,7 @@ impl BamlRuntime {
 
         let call_ctx = call_ctx.build();
 
-        let rt = bridge_cffi::engine::get_tokio_runtime().map_err(bridge_error_to_napi)?;
+        let rt = bridge_cffi::get_tokio_runtime().map_err(bridge_error_to_napi)?;
 
         // The whole Result -> BamlOutboundResult translation (incl. the
         // catch_unwind -> SdkPanic boundary and error/panic routing) lives in
@@ -97,7 +97,7 @@ impl BamlRuntime {
         collectors: Option<Vec<&Collector>>,
         abort_controller: Option<&AbortController>,
     ) -> napi::Result<PromiseRaw<'e, Buffer>> {
-        let runtime = bridge_cffi::engine::get_runtime().map_err(bridge_error_to_napi)?;
+        let runtime = bridge_cffi::get_runtime().map_err(bridge_error_to_napi)?;
         let kwargs = decode_args(args_proto.as_ref(), &function_name)?;
         let host_ctx = ctx.and_then(|c| c.host_span_context());
         let cancel = abort_controller
